menambahkan seeder untuk setiap model

+++ database/migrations/2023_07_02_171150_create_uraians_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('proyek_id')->constrained()->onDelete('cascade');
             $table->string('nama_uraian');
-            $table->integer('total_biaya')->nullable();
+            $table->float('total_biaya', 12, 2)->nullable();
             $table->timestamps();
         });
     }
